Update iptables.sh - Added DNS Rules for OpenVPN
Added for loop logic to iterate through name_server_list array and add iptables rules for output DNS requests on eth0. This will allow OpenVPN to look up its remote server hostname on connection failure and reconnect. This DNS request should only work when the OpenVPN default route is lost. It is a compromise to have OpenVPN reliably be able to reconnect.