Update start.sh - Updated resolve.conf logic
Defaulting to override of the existing resolve conf. Docker-compose adds an internal resolver at the top of the file when an bridge driver network is created. It's run on loopback 127.0.0.11 and used to resolve the other containers in that docker-compose network. Non-local queries are forwarded to the host upstream resolver. OpenVPN doesn't handle the fail over to secondary and tertiary DNS well on ping reset so it will cause repeated failures if the nameserver 127.0.0.11 statement isn't removed.
See automatic DNS resolution here: https://docs.docker.com/network/bridge/