Update iptables.sh - Relax Iptables Int Req
Added comments above the iptables inserts for DNS lookups. Discovered in testing that OpenVPN keeps the default route while it's in a failure and reconnect state after a server times out. DNS then fails repeatedly even with the rules in place for eth0. Removed the interface requirement for eth0 in case the request is originating from the tun0 or loopback interfaces in this failure state. Stopping iptables seems to resolve the issue, so this rule change should fix the problem. Last resort is to relax the return DNS rule in case the IP of the DNS server isn't what's expected because of NAT or other docker weirdness.